pyproject.toml
[project]
name = "shelfshift-demo"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "babel>=2.18.0",
    "fastapi>=0.135.1",
    "jinja2>=3.1.6",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "shelfshift>=1.0.2",
    "uvicorn>=0.41.0",
]

[project.scripts]
shelfshift-demo = "shelfshift_demo.main:run"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pandas>=3.0.1",
    "pytest>=9.0.2",
    "ruff>=0.15.4",
    "ty>=0.0.20",
]

[build-system]
requires = ["uv_build>=0.9.3,<0.10.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ty.environment]
python-version = "3.11"
